move: 잘못된 클래스 위치 이동
