+++ src/main/java/roomescape/reservation/client/PaymentClient.java
@@ -2,7 +2,7 @@ package roomescape.reservation.client;
 
 import org.springframework.http.ResponseEntity;
 
-import roomescape.reservation.service.PaymentApprovalRequest;
+import roomescape.reservation.dto.PaymentApprovalRequest;
 
 public interface PaymentClient {
